requirements.txt
filelock==3.15.1
fsspec==2024.6.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
opencv-python==4.10.0.84
pillow==10.3.0
sympy==1.12.1
torch==2.3.1
typing_extensions==4.12.2
